requirements.txt
mysql-connector-python~=8.0.24
pyyaml~=5.4.1
flask~=1.1.2